refactor: use container_hash/hash* over functional/hash*
The later just include the former.
